requirement.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mtcnn==1.0.0
namex==0.0.9
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.7
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pywin32==308
pyzmq==26.2.0
requests==2.32.3
rich==14.0.0
scipy==1.15.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
stack-data==0.6.3
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tornado==6.4.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
